Extract pubsub commands to mixin
